Many modifications to kd_tree code
Moved all the test code from kd_tree to new make_kd_tree.R file
Am coding a 2d test grid with some graphing to build and test the
ability to navigate the tree as well as graph the square regions the
cuts make, and to search outside of a leaf if the test point is near.
Also added a knn_single_point function that returns distance instead of
just the indices of the NN. Then tested that new function.