Added distances and other enhancements to knn code
moved library(scales) to top for consistency.
renamed somethings for clarity (e.g. single_point_train_grid to
train_grid to make it clear that it's not a single point, even if the
eventual test only tests knn using a single point on the grid.)

retained xy, which now is just a clone of train_grid for legacy code

Added if_necessary_convert_test_from_vector_to_tibble function to reduce
code duplication and use the function name to add clarity/context to somewhat
confusing code chunk.

added a features argument to the knn function so that the incoming train
set can be filtered to those features for the conversion of test to a
tibble. also features is used to filter to appropriate columns instead
of names(test).

knn now computes and stores the non-normalized distances, but it does
this using the returned indices of the knn_single_point function instead
of switching to use the knn_single_point_nn_indices_and_distances
function because we may want the normalization featuer only present in
the knn_single_point to get us the indices, even if we return the
distances as normalized.

adjsuted test code to use the new variable names and functions