Completed initial testing of KD trees.
Created a version of KD trees that simply alternates through the cut
variables (KDTreeSimple) and debugged the one I originally wrote which
tries to find the best cut (by minimizing variance under the constraint
that the leaves have a sizeable amount of the data in each.) Both will
cycle through variables and eventually reuse. Also created a 2d visual
test using a hand made data set. that confirms the code is working.

TODO: need to implement searching the KD tree for all the nearest
neighbors.